Add redirections for jsonnext and s
Add a jsonnet config file that is used to generated the firebase `json`
config file with HTTP redirections and the HTML files with HTML
redirections.

HTML redirections are configured for `jsonnext` and `s` modules. HTTP
redirections are configured just for `jsonnext`.

Update `Makefile` to run `jsonnet` to generate the config and deployed
files and to have the `deploy` target first generate the config.

Update `README.md` to document redirections.