doc: Document sw requirements, add make target to install
Document the software requirements to generate and deploy the
configuration and files for the redirections. Split the less necessary
information from the README into separate files, so the README has just
what you need to operate the repo quickly.

Add some targets to the `Makefile` to install these software
requirements for you (`software-update`, `firebase-update` and
`jsonnet-update`). Firebase is only updated if it is not the latest
version as it is a big download (120MB), so don't do it if we don't need
to.

Issue: #3