build: Add cloudbuild check and deployment
Update Makefile with `check-diff` and `deploy-on-master` targets.
`check-diff` ensures that generated configs are up-to-date.
`deploy-on-master` only deploys if `BRANCH_NAME` environment variable
is `master`.

Add `cloudbuild.yaml` file with `check-diff` and `deploy` steps.

We will manually set up a cloudbuild trigger on foxygoat-ab0f2 that will
run the cloudbuild on push to any branch.

link: https://console.cloud.google.com/cloud-build/triggers?project=foxygoat-ab0f2